+++ tests/general/default_config_test.go
@@ -58,11 +58,13 @@ func TestDefaultGatewayConfig(t *testing.T) {
 			config := collector.EffectiveConfig(t, 55554)
 			require.Equal(t, map[string]any{
 				"exporters": map[string]any{
-					"sapm": map[string]any{
-						"access_token": "<redacted>",
-						"endpoint":     "https://ingest.not.real.signalfx.com/v2/trace",
+					"otlphttp": map[string]any{
+						"traces_endpoint": "https://ingest.not.real.signalfx.com/v2/trace/otlp",
 						"sending_queue": map[string]any{
 							"num_consumers": 32,
+						},
+						"headers": map[string]any{
+							"X-SF-Token": "<redacted>",
 						},
 					},
 					"signalfx": map[string]any{
@@ -209,7 +211,7 @@ func TestDefaultGatewayConfig(t *testing.T) {
 							"receivers":  []any{"prometheus/internal"},
 						},
 						"traces": map[string]any{
-							"exporters":  []any{"sapm"},
+							"exporters":  []any{"otlphttp"},
 							"processors": []any{"memory_limiter", "batch"},
 							"receivers":  []any{"jaeger", "otlp", "sapm", "zipkin"},
 						},
@@ -256,15 +258,17 @@ func TestDefaultAgentConfig(t *testing.T) {
 					"debug": map[string]any{
 						"verbosity": "detailed",
 					},
-					"otlp": map[string]any{
+					"otlp/gateway": map[string]any{
 						"endpoint": ":4317",
 						"tls": map[string]any{
 							"insecure": true,
 						},
 					},
-					"sapm": map[string]any{
-						"access_token": "<redacted>",
-						"endpoint":     "https://ingest.not.real.signalfx.com/v2/trace",
+					"otlphttp": map[string]any{
+						"headers": map[string]any{
+							"X-SF-Token": "<redacted>",
+						},
+						"traces_endpoint": "https://ingest.not.real.signalfx.com/v2/trace/otlp",
 					},
 					"signalfx": map[string]any{
 						"access_token":       "<redacted>",
@@ -422,7 +426,7 @@ func TestDefaultAgentConfig(t *testing.T) {
 							"receivers":  []any{"prometheus/internal"},
 						},
 						"traces": map[string]any{
-							"exporters":  []any{"sapm", "signalfx"},
+							"exporters":  []any{"otlphttp", "signalfx"},
 							"processors": []any{"memory_limiter", "batch", "resourcedetection"},
 							"receivers":  []any{"jaeger", "otlp", "zipkin"},
 						},